+++ src/App.tsx
@@ -1,23 +1,15 @@
 import logo from './logo.svg'
 import './App.css'
-import { Account, NetworkType } from 'symbol-sdk'
+import CreateFromPrivateKey from './component/CreateFromPrivateKey'
+import GenerateNewAccount from './component/GenerateNewAccount'
 
 function App() {
-  const accountCreate = () => {
-    const account = Account.generateNewAccount(NetworkType.TEST_NET)
-    console.log(
-      'Your new account address is:',
-      account.address.pretty(),
-      'and its private key',
-      account.privateKey
-    )
-  }
-
   return (
     <div className="App">
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
-        <button onClick={accountCreate}>アカウントの作成</button>
+        <CreateFromPrivateKey></CreateFromPrivateKey>
+        <GenerateNewAccount></GenerateNewAccount>
       </header>
     </div>
   )
